+++ laplace_server/src/service/websocket.rs
@@ -2,11 +2,12 @@ use std::io;
 use std::ops::ControlFlow;
 use std::time::{Duration, Instant};
 
-use axum::extract::ws::{Message, WebSocket};
+use axum::extract::ws;
+use axum::extract::ws::WebSocket;
 use derive_more::From;
 use futures::stream::{SplitSink, SplitStream};
 use futures::{SinkExt, StreamExt};
-pub use laplace_wasm::route::websocket::Message as WsMessage;
+pub use laplace_wasm::route::websocket::{Message, MessageIn, MessageOut};
 use tokio::time;
 use truba::{Context, Sender, UnboundedMpscChannel};
 
@@ -21,7 +22,7 @@ enum WsError {
 }
 
 #[derive(Debug)]
-pub struct WsServiceMessage(pub WsMessage);
+pub struct WsServiceMessage(pub MessageOut);
 
 impl truba::Message for WsServiceMessage {
     type Channel = UnboundedMpscChannel<Self>;
@@ -34,7 +35,7 @@ pub struct WebSocketService {
     hb: Instant,
 
     lapp_service_sender: Sender<LappServiceMessage>,
-    ws_sender: SplitSink<WebSocket, Message>,
+    ws_sender: SplitSink<WebSocket, ws::Message>,
     ws_receiver: SplitStream<WebSocket>,
 }
 
@@ -73,12 +74,10 @@ impl WebSocketService {
                     }
                 },
                 Some(WsServiceMessage(msg)) = messages_in.recv() => {
-                    match msg {
-                        WsMessage::Text(text) => if self.handle_text(text).await.is_break() {
-                            break;
-                        },
+                    if self.handle_service_message(msg).await.is_break() {
+                        break;
                     }
-                }
+                },
             });
             self.close().await;
         });
@@ -92,20 +91,18 @@ impl WebSocketService {
         if Instant::now().duration_since(self.hb) > Self::CLIENT_TIMEOUT {
             // heartbeat timed out
             log::debug!("Websocket Client heartbeat failed, disconnecting!");
+            self.send_to_lapp(MessageIn::Timeout);
 
             // don't try to send a ping
-            return ControlFlow::Break(());
-        }
-
-        if let Err(err) = self.ws_sender.send(Message::Ping(Vec::new())).await {
-            log::error!("WS error: {err:?}");
+            ControlFlow::Break(())
+        } else if !self.send_to_ws(None, ws::Message::Ping(Vec::new())).await {
             ControlFlow::Break(())
         } else {
             ControlFlow::Continue(())
         }
     }
 
-    fn handle_ws_message(&mut self, msg: Result<Message, axum::Error>) -> ControlFlow<(), ()> {
+    fn handle_ws_message(&mut self, msg: Result<ws::Message, axum::Error>) -> ControlFlow<(), ()> {
         let msg = match msg {
             Ok(msg) => msg,
             Err(err) => {
@@ -115,36 +112,41 @@ impl WebSocketService {
         };
 
         match msg {
-            Message::Text(text) => {
-                log::debug!("Receive WS message: {text}");
-                if let Err(err) = self
-                    .lapp_service_sender
-                    .send(LappServiceMessage::Websocket(WsMessage::Text(text)))
-                {
-                    log::error!("Error occurs when send to lapp service: {err:?}");
-                }
+            ws::Message::Text(text) => {
+                log::debug!("Receive WS text: {text}");
+                self.send_to_lapp(Message::Text(text).into());
             },
-            Message::Binary(_bin) => {},
-            Message::Close(_close_frame) => {
+            ws::Message::Binary(bin) => {
+                log::debug!("Receive WS binary: {bin:?}");
+                self.send_to_lapp(Message::Binary(bin).into());
+            },
+            ws::Message::Close(close_frame) => {
+                log::debug!("Receive WS close: {close_frame:?}");
+                self.send_to_lapp(Message::Close.into());
                 return ControlFlow::Break(());
             },
 
-            Message::Pong(_) => {
+            ws::Message::Pong(_) => {
                 self.hb = Instant::now();
             },
             // You should never need to manually handle Message::Ping, as axum's websocket library
             // will do so for you automagically by replying with Pong and copying the v according to
             // spec. But if you need the contents of the pings you can see them here.
-            Message::Ping(_) => {
+            ws::Message::Ping(_) => {
                 self.hb = Instant::now();
             },
         }
         ControlFlow::Continue(())
     }
 
-    async fn handle_text(&mut self, text: String) -> ControlFlow<(), ()> {
-        if let Err(err) = self.ws_sender.send(Message::Text(text)).await {
-            log::error!("WS send error: {err:?}");
+    async fn handle_service_message(&mut self, MessageOut { id, msg }: MessageOut) -> ControlFlow<(), ()> {
+        let id = Some(id);
+        let sent = match msg {
+            Message::Text(text) => self.send_to_ws(id, ws::Message::Text(text)).await,
+            Message::Binary(text) => self.send_to_ws(id, ws::Message::Binary(text)).await,
+            Message::Close => self.send_to_ws(id, ws::Message::Close(None)).await,
+        };
+        if !sent {
             ControlFlow::Break(())
         } else {
             ControlFlow::Continue(())
@@ -152,8 +154,33 @@ impl WebSocketService {
     }
 
     async fn close(&mut self) {
-        if let Err(err) = self.ws_sender.send(Message::Close(None)).await {
-            log::error!("WS close send error: {err:?}");
+        self.ws_sender.send(ws::Message::Close(None)).await.ok();
+    }
+
+    async fn send_to_ws(&mut self, id: Option<String>, msg: ws::Message) -> bool {
+        let sent;
+        let result;
+
+        if let Err(err) = self.ws_sender.send(msg).await {
+            log::error!("WS send error: {err:?}");
+            result = Err(err.to_string());
+            sent = false;
+        } else {
+            result = Ok(());
+            sent = true;
+        }
+
+        if let Some(id) = id {
+            self.send_to_lapp(MessageIn::Response { id, result });
+        } else if let Err(err) = result {
+            self.send_to_lapp(MessageIn::Error(err.to_string()));
+        }
+        sent
+    }
+
+    fn send_to_lapp(&self, msg: MessageIn) {
+        if let Err(err) = self.lapp_service_sender.send(LappServiceMessage::WebSocket(msg)) {
+            log::error!("Error occurs when send to lapp service: {err:?}");
         }
     }
 }